+++ src/object.js
@@ -203,3 +203,19 @@ export let mergeOverAllWith = _.curryN(3, (customizer, fns, ...x) =>
 
 // ([f, g]) -> (x, y) -> {...f(x, y), ...g(x, y)}
 export let mergeOverAllArrays = mergeOverAllWith(mergeArrays)
+
+// (x -> y) -> k -> {k: x} -> y
+export let getWith = _.curry((customizer, path, object) =>
+  customizer(_.get(path, object))
+)
+
+// ({a} -> {b}) -> {a} -> {a, b}
+export let expandObject = _.curry((transform, obj) => ({
+  ...obj,
+  ...transform(obj),
+}))
+
+// k -> (a -> {b}) -> {k: a} -> {a, b}
+export let expandObjectBy = _.curry((key, fn, obj) =>
+  expandObject(getWith(fn, key))(obj)
+)