+++ CHANGELOG.md
@@ -1,6 +1,12 @@
-# 1.61.1
+﻿# 1.63.0
 - Add partial currying support to `mergeOverAll`
 - Add `mergeOverAllWith` and `mergeOverAllArrays`
+
+# 1.62.1
+- Better currying for `logic` methods (`ifElse`, `when`, `whenExists`, etc)
+
+# 1.62.0
+- Add `getWith`, `expandObject`, and `expandObjectBy`
 
 # 1.61.0
 - Add `compactMap`